Cambios en FechaInicioService y EscaleraProduccionService para el calculo de escalera.

See merge request oilsoftdev/oportexp-economic-eva!19

+++ src/main/java/mx/com/innovating/cloud/data/models/CalculoNumPozosResult.java
@@ -9,7 +9,7 @@ import lombok.*;
 @NoArgsConstructor
 public class CalculoNumPozosResult {
 
-    private Integer nPozos;
+    private Double nPozos;
     private String tipoCalculo;
 
 }
